[fix] DTO Request, Response 명칭 수정, Service, ServiceImpl 구조 수정
